Enable building arm64 if requested
Add arm64 to the list of architectures to build when requested by the
user, this can be done by setting APPLE_SILICON_SUPPORT environment or
Make variable to "yes", for example:

   make APPLE_SILICON_SUPPORT=yes

We intend to provide arm64 builds in the default package, but until it
is available in stable Xcode releases, arm64 capability will be under
the feature gate to avoid breaking stable Xcode builds. Currently using
APPLE_SILICON_SUPPORT requires Xcode 12+ beta to be active.

Also note that currently only the macOS build is enabled. It is possible
to build arm64 for iOS Simulator, but it's currently not possible to
package the results into the same fat binary along with the arm64 slice
for the actual iOS devices. This issue will be solved later. It seems
that using XCFrameworks should allow mixing targets the way we need it.