Differentiate iOS Simulator headers
New arm64 architecture for iOS Simulator is conflicting with arm64
architecture for the real iOS devices. Though the architecture is the
same, the target is not. A different SDK is used to build arm64 for
Apple Silicon devices, and it's not compatible in the strict sense.

OpenSSL currently has per-architecture headers with platform-specific
details. Make sure that iOS ones don't conflict with iOS Simulator by
making iOS use "openssl_ios_arm64.h" and iOS Simulator will be using
"openssl_ios_sim_arm64.h"