Extra package version digits
Teenage mutant ninja turtles
...sorry.

Occasionally we need to publish the same OpenSSL version several times,
such as when the version did not really change but Apple platforms did.
Normally you increment the "patch version" for this, or some dedicated
"package version", but we don't have that luxury with OpenSSL. We must
keep the version strictly semver, and all the digits are already taken
with OpenSSL version, even with some extra on top.

The current latest version of OpenSSL is 1.1.1h. However, semver does
not allow letters so we transform that into 1.1.108:

   1.1.1h => 1.1.(1 * 100) + ('h' - 'a' + 1) => 1.1.108

In the similar vein, add yet another 'package version' part to this
whole mess:

   1.1.1h-1 => 1.1.((1 * 100) + ('h' - 'a' + 1)) * 100 + 1 => 1.1.10801

It's not the prettiest solution but... doin' what I can with what I got.
This approach is semver-compatible, pleasing our fruit company overlords.
The resulting version ordering is useful. Since the 1.1.107 has been the
only release so far, migration to 1.1.10701 is smooth. The transformation
is reversible and unambiguous (well, let's hope we won't need to release
more than 100 versions of the same OpenSSL).

The source of truth for CLOpenSSL is the PACKAGE_VERSION variable in the
Makefile. Bump it to release a new package version. Reset it to "1" when
updating OpenSSL version.

Note that the tag names are using the same approach because they are
significant to Carthage and CocoaPods.